$load and reload shouldn't run on background thread
We modified $show logic so that it triggers the attached actions on the background thread, because otherwise the everytime a user switches between tabs the entire $show action call chain would execute first, and then go on to render the view.

However we don't need to (and should not) do the same for $load. First of all $load is not related to tab switching, and having all these different threads is not good.

Need to minimize thread usage as much as possible because threading introduces timing glitches.

Also got rid of explicitly running reload on the background thread because it needlessly complicated things. Again, better to have as few threads as possible