Goodbye Russian Doll
RussianDollView was a very very early design decision while exploring different options for modularity. This is not relevant anymore since it's become clear how Jasonette will approach modularity moving forward, and it won't involve RussianDollView.

It's been just sitting there and only making things harder to debug and write code, and making the code much more complex than it needs to be.

Getting rid of it.