Multithreading Solution
# Problem

In some cases a view listens to multiple events. And sometimes multiple events can fire simultaneously, resulting in Jasonette not knowing what event to trigger.

This has not been that problematic since there weren't that many events to listen to, but with all the new feature additions, there can be infinite number of events to listen to. This is problematic because currently Jasonette is intentionally single threaded (in Jasonette action call chain sense).

For example, the app may try to `$render` on `$load` event, but also set up a websocket session and listen to all incoming websocket messages which also will call a `$render` in the end. If you get 100 websocket messages before the `$render` from the `$load` event has been fully finished, we need a way to process this.

# Failed Attempts

I have tried two approaches:

1. When there's an action call chain executing when a new event comes in, IGNORE the new event
2. When there's an action call chain executing when a new event comes in, OVERRIDE the currently executing chain with the new event call, thereby abandoning the previous call chain.

Obviously, none of the two approaches are sound.

# Final Solution

Instead of dropping or overriding with the new event, now there's an event queue, which pushes all incoming events. This doesn't mean action call chains are now run multithreaded, because that would result in even more confusing behaviors (For example one call chain may be trying to render the template with certain state, while another call chain is trying to render with a different state. Jasonette shouldn't try to run those simultaneously since it will result in unexpected render behavior. BTW this was why Jasonette action call chains are singl threaded in the first place)

So here's the final, simple solution that works:

1. Push all incoming events into an event queue
2. Pop and run every time a call chain finishes
3. Keep repeating until the event queue is empty